Move tokio_block_on
